requirements.txt
altgraph==0.17.4
asgiref==3.7.2
astroid==3.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.40.0
colorama==0.4.6
crispy-bootstrap5==2023.10
cryptography==42.0.0
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
Django==4.2.13
django-allauth==0.60.1
django-cleanup==8.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
filelock==3.13.1
graphviz==0.20.1
gunicorn==22.0.0
idna==3.6
isort==5.13.2
mccabe==0.7.0
numpy==1.26.2
oauthlib==3.2.2
packaging==24.0
pandas==2.1.3
pefile==2023.2.7
pillow==10.2.0
pipenv==2023.10.24
platformdirs==3.11.0
psycopg2==2.9.9
pycparser==2.21
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
PyJWT==2.8.0
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.24.6
whitenoise==6.6.0
xmltodict==0.13.0
